class-projects/Big Agents/Travel Designer Agent/pyproject.toml
[project]
name = "ai-travel-designer-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colorama>=0.4.6",
    "openai-agents>=0.0.19",
    "python-dotenv>=1.1.0",
]
